requirements.txt
# Core Data Science
numpy
pandas

# Machine Learning
scikit-learn
torch

# Visualization
matplotlib
seaborn

# Utilities
joblib
pyyaml

# Statistical Analysis
scipy